requirements.txt
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
oauthlib==3.2.2
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.30.0
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.4.1
tweepy==4.14.0
urllib3==2.0.2
